:truck: Create test_command.py
